# Lazily import `torch/pydantic` in `json` module, speedup `from monty.json import` by 10x