### Summary

- Lazily import `torch/pydantic` for `monty.json`, thanks for @janosh's profiling in https://github.com/janosh/pymatviz/issues/209#issuecomment-2423174822, would significantly speed up `from monty.json import MSONable/MontyEncoder/MontyDecoder/...`

---

## Performance comparison

- Following benchmark done with **optional dependencies installed** (otherwise import would be skipped), averaged across 10 runs
- Python 3.12.7, MacOS 15.0.1 with M3 chip
- Profile with: `python -X importtime -c  "from monty.json import MSONable" 2> monty.log && tuna monty.log`

<details>
<summary> Benchmark Script (by GPT) </summary>

```python
import time
import subprocess


def measure_import_time(command: str, repeats: int):
    total_time = 0

    for _ in range(repeats):
        start_time = time.time()

        # Run the Python command in a new process
        subprocess.run(["python", "-c", command], stdout=subprocess.PIPE, stderr=subprocess.PIPE)

        end_time = time.time()
        total_time += (end_time - start_time)

    avg_time_ms = (total_time / repeats) * 1000
    print(f"Average import time for '{command}': {avg_time_ms:.3f} ms")

def main():
    commands = [
        "from monty.json import MSONable",
        "from monty.json import MontyEncoder",
        "from monty.json import MontyDecoder",
    ]

    for command in commands:
        measure_import_time(command, repeats=10)

if __name__ == "__main__":
    main()

```

</details>


### Import time from master branch
```
Average import time for 'from monty.json import MSONable': 924.585 ms
Average import time for 'from monty.json import MontyEncoder': 735.797 ms
Average import time for 'from monty.json import MontyDecoder': 720.402 ms
```

Import time traced for `from monty.json import MSONable`:
<img width="1505" alt="image" src="https://github.com/user-attachments/assets/984d1cb3-7060-46bb-a694-0fb07c04dbc0">

### Import time from tip of this PR
```
Average import time for 'from monty.json import MSONable': 109.053 ms
Average import time for 'from monty.json import MontyEncoder': 111.683 ms
Average import time for 'from monty.json import MontyDecoder': 103.302 ms
```

<img width="1356" alt="image" src="https://github.com/user-attachments/assets/1c8e7ee0-6d0e-4311-be67-fc8207f3ddc3">

---

### After lazily import `torch`

```
Average import time for 'from monty.json import MSONable': 122.606 ms
Average import time for 'from monty.json import MontyEncoder': 123.456 ms
Average import time for 'from monty.json import MontyDecoder': 117.621 ms
```

<img width="1505" alt="image" src="https://github.com/user-attachments/assets/59bd3af4-1a99-4b73-9ca1-c93b3a62d4b1">

---

### After lazily import `numpy` as well (reverted, as use of np is very common)

```
Average import time for 'from monty.json import MSONable': 58.586 ms
Average import time for 'from monty.json import MontyEncoder': 59.323 ms
Average import time for 'from monty.json import MontyDecoder': 58.300 ms
```

<img width="1505" alt="image" src="https://github.com/user-attachments/assets/b1de8b21-cc9e-4d48-8c31-a518575a96e3">

---

### After lazily import `pydantic` too (limited improvement from here)

```
Average import time for 'from monty.json import MSONable': 54.003 ms
Average import time for 'from monty.json import MontyEncoder': 53.038 ms
Average import time for 'from monty.json import MontyDecoder': 52.875 ms
```

<img width="1505" alt="image" src="https://github.com/user-attachments/assets/6566c8f3-f038-4676-b296-3f7c6cc9f8ab">

---

### After lazily import `ruamel.yaml` (reverted, no obvious improvement)

```
Average import time for 'from monty.json import MSONable': 52.769 ms
Average import time for 'from monty.json import MontyEncoder': 52.191 ms
Average import time for 'from monty.json import MontyDecoder': 60.379 ms
```

<img width="1505" alt="image" src="https://github.com/user-attachments/assets/35c347b9-afbc-4fe4-939a-64a607b38538">



<!-- This is an auto-generated comment: release notes by coderabbit.ai -->
## Summary by CodeRabbit

- **New Features**
	- Enhanced type safety and error handling in JSON processing.
- **Chores**
	- Updated GitHub Actions workflow for linting to use the latest Python setup action.
	- Modified pre-commit hook configuration to run codespell during the pre-commit phase.
	- Updated testing workflow to include Python 3.12 in the testing environment.
<!-- end of auto-generated comment: release notes by coderabbit.ai -->