@DanielYang59 2024-10-21
Cool thanks for the comment! Sure thing I would revert the part for `numpy`, can you please review #714 which would unblock this PR (otherwise nest those import block would make it really hard to see what optional dependency is required)? Really appreciate that